Factor out project.util.django_graphql_session_forms.py.
